+++ backend/globals.py
@@ -1,0 +1,1 @@
+initial_admin_setup_needed = False